requirements.txt
opencv-python
numpy
tensorflow
streamlit
transformers
torch
torchvision
pillow